+++ lib/models.py
@@ -87,6 +87,10 @@ class MetatraderPosition(TypedDict):
     """Profit of the already closed part, including commissions and swap."""
     commission: Optional[float]
     """Optional position commission."""
+    reason: str
+    """Position opening reason. One of POSITION_REASON_CLIENT, POSITION_REASON_EXPERT, POSITION_REASON_MOBILE,
+    POSITION_REASON_WEB, POSITION_REASON_UNKNOWN. See
+    https://www.mql5.com/en/docs/constants/tradingconstants/positionproperties#enum_position_reason"""
 
 
 class MetatraderOrder(TypedDict):
@@ -141,6 +145,10 @@ class MetatraderOrder(TypedDict):
     updatePending: Optional[bool]
     """Optional flag indicating that order client id and original comment was not
     identified yet and will be updated in a future synchronization packet."""
+    reason: str
+    """Order opening reason. One of ORDER_REASON_CLIENT, ORDER_REASON_MOBILE, ORDER_REASON_WEB,
+    ORDER_REASON_EXPERT, ORDER_REASON_SL, ORDER_REASON_TP, ORDER_REASON_SO, ORDER_REASON_UNKNOWN. See
+    https://www.mql5.com/en/docs/constants/tradingconstants/orderproperties#enum_order_reason."""
 
 
 class MetatraderHistoryOrders(TypedDict):
@@ -203,6 +211,11 @@ class MetatraderDeal(TypedDict):
     updatePending: Optional[bool]
     """Optional flag indicating that deal client id and original comment was not
     identified yet and will be updated in a future synchronization packet"""
+    reason: Optional[str]
+    """Optional deal execution reason. One of DEAL_REASON_CLIENT, DEAL_REASON_MOBILE, DEAL_REASON_WEB,
+    DEAL_REASON_EXPERT, DEAL_REASON_SL, DEAL_REASON_TP, DEAL_REASON_SO, DEAL_REASON_ROLLOVER, DEAL_REASON_VMARGIN,
+    DEAL_REASON_SPLIT, DEAL_REASON_UNKNOWN. See
+    https://www.mql5.com/en/docs/constants/tradingconstants/dealproperties#enum_deal_reason."""
 
 
 class MetatraderDeals(TypedDict):